Added possibility to move backwards.
Changes in Piece class:
If a human player has enough spaces to move in both directions, he/she can choose which way to go by pressing the F (forward) or B (backward) button which appear on the screen.
If there aren't enough spaces to move in one of those directions, the opposite direction is automatically chosen.
CPU players always move forward, unless there are not enough spaces to do so. In that case they move backwards.